+++ config/configtelemetry/configtelemetry.go
@@ -38,6 +38,8 @@ const (
 	metricsLevelCfg = "metrics-level"
 )
 
+const UseOpenTelemetryForInternalMetrics = false
+
 var metricsLevelPtr = new(Level)
 
 // Flags is a helper function to add telemetry config flags to the service that exposes