+++ internal/testcomponents/example_factories.go
@@ -15,8 +15,9 @@
 package testcomponents
 
 import (
+	"go.uber.org/multierr"
+
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/consumer/consumererror"
 	"go.opentelemetry.io/collector/exporter/loggingexporter"
 	"go.opentelemetry.io/collector/exporter/otlpexporter"
 	"go.opentelemetry.io/collector/exporter/otlphttpexporter"
@@ -54,39 +55,31 @@ func DefaultComponents() (
 	component.Factories,
 	error,
 ) {
-	var errs []error
+	var errs error
 
 	extensions, err := component.MakeExtensionFactoryMap(
 		zpagesextension.NewFactory(),
 		ballastextension.NewFactory(),
 	)
-	if err != nil {
-		errs = append(errs, err)
-	}
+	errs = multierr.Append(errs, err)
 
 	receivers, err := component.MakeReceiverFactoryMap(
 		otlpreceiver.NewFactory(),
 	)
-	if err != nil {
-		errs = append(errs, err)
-	}
+	errs = multierr.Append(errs, err)
 
 	exporters, err := component.MakeExporterFactoryMap(
 		loggingexporter.NewFactory(),
 		otlpexporter.NewFactory(),
 		otlphttpexporter.NewFactory(),
 	)
-	if err != nil {
-		errs = append(errs, err)
-	}
+	errs = multierr.Append(errs, err)
 
 	processors, err := component.MakeProcessorFactoryMap(
 		batchprocessor.NewFactory(),
 		memorylimiterprocessor.NewFactory(),
 	)
-	if err != nil {
-		errs = append(errs, err)
-	}
+	errs = multierr.Append(errs, err)
 
 	factories := component.Factories{
 		Extensions: extensions,
@@ -95,5 +88,5 @@ func DefaultComponents() (
 		Exporters:  exporters,
 	}
 
-	return factories, consumererror.Combine(errs)
+	return factories, errs
 }