+++ receiver/otlpreceiver/otlp_test.go
@@ -537,7 +537,7 @@ func TestOTLPReceiverTrace_HandleNextConsumerResponse(t *testing.T) {
 
 				require.Equal(t, tt.expectedReceivedBatches, len(sink.AllTraces()))
 
-				obsreporttest.CheckReceiverTraces(t, config.NewIDWithName(typeStr, exporter.receiverTag), "grpc", int64(tt.expectedReceivedBatches), int64(tt.expectedIngestionBlockedRPCs))
+				require.NoError(t, obsreporttest.CheckReceiverTraces(config.NewIDWithName(typeStr, exporter.receiverTag), "grpc", int64(tt.expectedReceivedBatches), int64(tt.expectedIngestionBlockedRPCs)))
 			})
 		}
 	}
@@ -572,6 +572,44 @@ func TestGRPCInvalidTLSCredentials(t *testing.T) {
 	assert.EqualError(t,
 		r.Start(context.Background(), componenttest.NewNopHost()),
 		`failed to load TLS config: for auth via TLS, either both certificate and key must be supplied, or neither`)
+}
+
+func TestGRPCMaxRecvSize(t *testing.T) {
+	addr := testutil.GetAvailableLocalAddress(t)
+	sink := new(consumertest.TracesSink)
+
+	factory := NewFactory()
+	cfg := factory.CreateDefaultConfig().(*Config)
+	cfg.GRPC.NetAddr.Endpoint = addr
+	cfg.HTTP = nil
+	ocr := newReceiver(t, factory, cfg, sink, nil)
+
+	require.NotNil(t, ocr)
+	require.NoError(t, ocr.Start(context.Background(), componenttest.NewNopHost()))
+
+	cc, err := grpc.Dial(addr, grpc.WithInsecure(), grpc.WithBlock())
+	require.NoError(t, err)
+
+	td := testdata.GenerateTracesManySpansSameResource(500000)
+	require.Error(t, exportTraces(cc, td))
+	cc.Close()
+	require.NoError(t, ocr.Shutdown(context.Background()))
+
+	cfg.GRPC.MaxRecvMsgSizeMiB = 100
+	ocr = newReceiver(t, factory, cfg, sink, nil)
+
+	require.NotNil(t, ocr)
+	require.NoError(t, ocr.Start(context.Background(), componenttest.NewNopHost()))
+	t.Cleanup(func() { require.NoError(t, ocr.Shutdown(context.Background())) })
+
+	cc, err = grpc.Dial(addr, grpc.WithInsecure(), grpc.WithBlock())
+	require.NoError(t, err)
+	defer cc.Close()
+
+	td = testdata.GenerateTracesManySpansSameResource(500000)
+	require.NoError(t, exportTraces(cc, td))
+	require.Len(t, sink.AllTraces(), 1)
+	assert.Equal(t, td, sink.AllTraces()[0])
 }
 
 func TestHTTPInvalidTLSCredentials(t *testing.T) {