+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 
 ## 🛑 Breaking changes 🛑
 
-- Remove `defaultcomponents` from core (#4087). 
-- Move configcheck.ValidateConfigFromFactories as internal function in service package (#3876).
+- Move configcheck.ValidateConfigFromFactories as internal function in service package (#3876)
+- Rename `configparser.Parser` as `config.Map` (#4075)
+- Rename `component.DefaultBuildInfo()` to `component.NewDefaultBuildInfo()` (#4129)
+- Remove `defaultcomponents` from core (#4087).
+
+## 💡 Enhancements 💡
+
+- Add Gen dependabot into CI (#4083)
 
 ## v0.36.0 Beta
 
@@ -19,6 +25,7 @@
 - Remove squash on `configtls.TLSClientSetting` and move TLS client configs under `tls` (#4063)
 - Rename TLS server config `*configtls.TLSServerSetting` from `tls_settings` to `tls` (#4063)
 - Split `service.Collector` from the `cobra.Command` (#4074)
+- Rename `memorylimiter` to `memorylimiterprocessor` (#4064)
 
 ## 💡 Enhancements 💡